+++ CHANGELOG.md
@@ -25,7 +25,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## Log
 
-### [0.6.0]
+### [0.7.0] - 2020-09-25
+
+#### Added
+
+* Open selection(s) in new editor
+
+#### Fixed
+
+* Fixed a bug where multiline random inserts were failing if the command needed to prompt the user for a choise. The affected commands were PERSON_NAME, COLOR, PARAGRAPH and HASH
+
+### [0.6.0] - 2020-09-24
 
 #### Updated